feat(asm): show hint for top element in stack info, show new styled stack info in doc and completion
Fixes #182
